+++ examples/using-emotion/src/pages/index.js
@@ -1,7 +1,7 @@
 import React, { Fragment } from "react"
 import { Helmet } from "react-helmet"
 import styled from "@emotion/styled"
-import { css, Global } from "@emotion/core"
+import { css, Global } from "@emotion/react"
 
 // Emotion supports different styling options, all of which are supported by gatsby-plugin-emotion out of the box